# ENH: Improve performance of finfo and _commonType