* The `finfo` contains a cache for dtypes, but the `np.complex128` dtype does not end up in the cache. The reason is that the `np.complex128` is converted to `np.float64` which is in the cache.

Performance improvement for `finfo(np.complex128)`:
```
Main: 2.07 µs ± 75 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
Pr: 324 ns ± 28.9 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

* Improve performance of `finfo` by making the cache check the first action in the `__new__`
* Improve performance of `_commonType` by re-using the expression for `a.dtype.type` and eliminating variables

The `finfo` and `_commonType` was part of the computatation time in `lstsq` when using scikit-rf. Since these methods are used in various other methods performance can improve there slightly as well.

<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      https://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      https://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      https://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->