@seberg 2023-01-26
Seems OK, but silly question: should this just be `@functools.lru_cache` and be done with it?  It enforces hashability, but the current code looks like it does so as well.  The 128 entry default seems plenty to me also.