@eendebakpt 2023-01-26
> Seems OK, but silly question: should this just be `@functools.lru_cache` and be done with it? It enforces hashability, but the current code looks like it does so as well. The 128 entry default seems plenty to me also.

Initially I rejected this idea because `finfo(1.1)` is allowed, and there are too many floats to fit into the original cache. For for an `lru_cache` with finite size this is no problem, so I now think this is a good idea. I created #23108 with an implementation and notes.